fix(compose,material): the static analysis queue for the drawable components and the material library — bit patterns in unsigned arithmetic with the math untouched, colour helpers and palette tables that cannot throw at static initialisation, optionals checked before they are read, integer quotients named for what they are, descriptions and verbs taken by reference, dead stores removed, and every finding that stands answered in place with its reason

+++ apps/spell-circle-canvas/src/common/compose/gallery/OrnamentKit.h
@@ -401,7 +401,7 @@ inline std::function<SkPath(SkSize)> starburstOutline(int spikes, float depth) {
     const int n = spikes * 2;
     for (int i = 0; i < n; ++i) {
       const float a = (float)i / (float)n * 6.28318f - 1.5708f;
-      const float k = (i & 1) ? 1.0f - depth : 1.0f;
+      const float k = ((unsigned)i & 1u) ? 1.0f - depth : 1.0f;
       const SkPoint pt = {cx + std::cos(a) * cx * k, cy + std::sin(a) * cy * k};
       if (i == 0)
         b.moveTo(pt);
